Unit Mvt. Update
Adds BoardGameManager which maintains an array of all tiles and which tile is selected.
Adds functionality to allow the basic unit to move to a selected tile, permitting it fits within speed and tile requirements (target tile cannot be mtn or water).
KNOWN ISSUE: see UnitControl.cs line 62 for more info.